Data races!
Needs to be able to stay in two different dir at the same
time. Now the program changes to the same dir becuase of value overwrite. What
to do?

fork() will create a new instance and discarde the data races and the change
dir problem (I think), but this also makes the problem impossible to regulate
through mutexes to the log file.

pthreads will have the chdir problem but will be able to utalize mutexes for
the write function. Don't think the chdir problem can be solved with mutexes.

Signed-off-by: mattemagikern <mansgariusson@gmail.com>